Refactor parse_udf
