+++ bigquery_etl/udf/publish_udfs.py
@@ -8,7 +8,7 @@ from google.cloud import bigquery
 from google.cloud import storage
 
 from bigquery_etl.util import standard_args
-from bigquery_etl.parse_udf import read_udf_dirs, accumulate_dependencies
+from bigquery_etl.udf.parse_udf import read_udf_dirs, accumulate_dependencies
 
 DEFAULT_PROJECT_ID = "moz-fx-data-shared-prod"
 DEFAULT_UDF_DIR = ["udf/", "udf_js/"]
