+++ tests/assert/all_fields_null/metadata.yaml
@@ -1,0 +1,2 @@
+description: ''
+friendly_name: All Fields Null
