+++ tests/assert/all_fields_null/udf.sql
