+++ tests/assert/approx_equals/udf.sql
