+++ tests/assert/array_empty/metadata.yaml
@@ -1,0 +1,2 @@
+description: ''
+friendly_name: Array Empty
