+++ tests/assert/array_empty/udf.sql
