+++ tests/assert/array_equals/udf.sql
