+++ tests/assert/array_equals_any_order/udf.sql
