+++ tests/assert/equals/udf.sql
