+++ tests/assert/error/udf.sql
