+++ tests/assert/histogram_equals/udf.sql
