+++ tests/assert/map_equals/metadata.yaml
@@ -1,0 +1,2 @@
+description: ''
+friendly_name: Map Equals
