+++ tests/assert/map_equals/udf.sql
