+++ tests/assert/null/metadata.yaml
@@ -1,0 +1,2 @@
+description: ''
+friendly_name: 'Null'
