+++ tests/assert/null/udf.sql
