+++ tests/assert/true/metadata.yaml
@@ -1,0 +1,2 @@
+description: ''
+friendly_name: 'True'
